fixes per PR comments 2/2
