+++ packages/shared/src/lib/core/contact/classes/index.ts
