+++ packages/shared/src/lib/core/contact/classes/tests/contact-manager.test.ts
