+++ packages/shared/src/lib/core/contact/index.ts
@@ -1,0 +1,3 @@
+export * from './classes'
+export * from './interfaces'
+export * from './stores'
