+++ packages/shared/src/lib/core/contact/interfaces/contact-address.interface.ts
@@ -6,4 +6,5 @@ export interface IContactAddress {
     address: string
     addressName: string
     contactId: string
+    networkId: string
 }
