+++ packages/shared/src/lib/core/contact/interfaces/contact-metadata.interface.ts
