+++ packages/shared/src/lib/core/contact/interfaces/network-contact-address-map.interface.ts
