+++ packages/shared/src/lib/core/contact/interfaces/network-contact-addresses-map.interface.ts
