+++ packages/shared/src/lib/core/contact/stores/index.ts
@@ -1,0 +1,2 @@
+export * from './selected-contact.store'
+export * from './selected-contact-network-id.store'
