+++ packages/shared/src/lib/core/contact/stores/selected-contact-network-id.store.ts
@@ -1,0 +1,17 @@
+import { get, writable } from 'svelte/store'
+
+export const selectedContactNetworkId = writable<string | undefined>(undefined)
+
+export function clearSelectedContactNetworkId(): void {
+    selectedContactNetworkId.set(undefined)
+}
+
+export function getSelectedContactNetworkId(): string | undefined {
+    return get(selectedContactNetworkId)
+}
+
+export function setSelectedContactNetworkId(networkId: string): void {
+    if (networkId) {
+        selectedContactNetworkId.set(networkId)
+    }
+}
