--- packages/shared/src/lib/core/contacts/index.ts
@@ -1,6 +1,0 @@
-// export * from './actions'
-export * from './classes'
-// export * from './helpers'
-export * from './interfaces'
-export * from './stores'
-// export * from './types'
