--- packages/shared/src/lib/core/contacts/stores/index.ts
@@ -1,2 +1,0 @@
-export * from './selected-contact.store'
-export * from './selected-network-id.store'
