--- packages/shared/src/lib/core/contacts/stores/selected-network-id.store.ts
@@ -1,17 +1,0 @@
-import { get, writable } from 'svelte/store'
-
-export const selectedNetworkId = writable<string | undefined>(undefined)
-
-export function clearSelectedNetworkId(): void {
-    selectedNetworkId.set(undefined)
-}
-
-export function getSelectedNetworkId(): string | undefined {
-    return get(selectedNetworkId)
-}
-
-export function setSelectedNetworkId(networkId: string): void {
-    if (networkId) {
-        selectedNetworkId.set(networkId)
-    }
-}
