+++ packages/shared/src/lib/core/profile/interfaces/persisted-profile.interface.ts
@@ -4,7 +4,7 @@ import { INft } from '@core/nfts'
 import { StrongholdVersion } from '@core/stronghold/enums'
 import { ProfileType } from '../enums'
 import { IProfileSettings } from './profile-settings.interface'
-import { IContactMap, INetworkContactAddressMap } from '@core/contacts/interfaces'
+import { IContactMap, INetworkContactAddressMap } from '@core/contact/interfaces'
 
 export interface IPersistedProfile {
     id: string
